added to front desk and moved methods from room there
