+++ lib/reservation.rb
@@ -4,7 +4,7 @@ require 'room'
 module Hotel
   class Reservation
     attr_reader :start_day, :end_day, :room, :id
-    @@reservations = []
+    @reservations = []
     def initialize(room, start_day, end_day)
       @room = room
       @start_day = Date.parse(start_day)
@@ -24,22 +24,6 @@ module Hotel
 
     def total_cost
       @total_nights * 200
-    end
-#all below needs tests
-    def create_reservation(room,start_day,end_day)
-      new_reservation = Hotel::Reservation.new(room,start_day,end_day)
-      @@reservations << new_reservation
-      store_res_by_room(new_reservation)
-      return new_reservation
-    end
-
-    def store_res_by_room(reservation)
-      room = find_room(reservation.room_number)
-      room.reservations << reservation
-    end
-
-    def find_res_by_date(date)
-      @@reservations.select{|reservation| reservation.start_day <= date && reservation.end_day >= date}
     end
   end
 end
